+++ src/commands/install/install-package.ts
@@ -1,0 +1,199 @@
+import chalk from "chalk";
+import {
+  getConfig,
+  getRuleSource,
+  getOriginalPresetPath,
+} from "../../utils/config";
+import { detectRuleType } from "../../utils/rule-detector";
+import { NormalizedConfig } from "../../types";
+import { withWorkingDirectory } from "../../utils/working-directory";
+import {
+  collectLocalRule,
+  collectNpmRule,
+  initRuleCollection,
+  addRuleToCollection,
+} from "../../utils/rule-collector";
+import { writeRulesToTargets } from "../../utils/rule-writer";
+import { writeMcpServersToTargets } from "../../utils/mcp-writer";
+import { expandRulesGlobPatterns } from "../../utils/glob-handler";
+
+/**
+ * Options for the install functions
+ */
+export interface InstallOptions {
+  /**
+   * Base directory to use instead of process.cwd()
+   */
+  cwd?: string;
+  /**
+   * Custom config object to use instead of loading from file
+   */
+  config?: NormalizedConfig;
+  /**
+   * allow installation on CI environments
+   */
+  installOnCI?: boolean;
+  /**
+   * Show verbose output during installation
+   */
+  verbose?: boolean;
+}
+
+/**
+ * Result of the install operation
+ */
+export interface InstallResult {
+  /**
+   * Whether the operation was successful
+   */
+  success: boolean;
+  /**
+   * Error message if the operation failed
+   */
+  error?: string;
+  /**
+   * Error stack trace for debugging (when available)
+   */
+  errorStack?: string;
+  /**
+   * Number of rules installed
+   */
+  installedRuleCount: number;
+  /**
+   * Number of packages installed
+   */
+  packagesCount: number;
+}
+
+/**
+ * Install rules for a single package (used within workspaces and standalone installs)
+ * @param options Install options
+ * @returns Result of the install operation
+ */
+export async function installPackage(
+  options: InstallOptions = {},
+): Promise<InstallResult> {
+  const cwd = options.cwd || process.cwd();
+
+  return withWorkingDirectory(cwd, async () => {
+    const config = options.config || getConfig();
+
+    const ruleCollection = initRuleCollection();
+
+    if (!config) {
+      return {
+        success: false,
+        error: "Configuration file not found",
+        installedRuleCount: 0,
+        packagesCount: 0,
+      };
+    }
+
+    // Check if rules are defined (either directly or through presets)
+    if (!config.rules || Object.keys(config.rules).length === 0) {
+      // If there are no presets defined either, show a message
+      if (!config.presets || config.presets.length === 0) {
+        return {
+          success: false,
+          error: "No rules defined in configuration",
+          installedRuleCount: 0,
+          packagesCount: 0,
+        };
+      }
+    }
+
+    let expandedRules: Record<string, string>;
+
+    try {
+      const expansion = await expandRulesGlobPatterns(config.rules, cwd);
+      expandedRules = expansion.expandedRules;
+
+      if (options.verbose) {
+        for (const [expandedKey, originalPattern] of Object.entries(
+          expansion.globSources,
+        )) {
+          console.log(
+            chalk.gray(`  Pattern "${originalPattern}" → ${expandedKey}`),
+          );
+        }
+      }
+    } catch (error) {
+      const errorMessage = `Error expanding glob patterns: ${error instanceof Error ? error.message : String(error)}`;
+
+      return {
+        success: false,
+        error: errorMessage,
+        errorStack: error instanceof Error ? error.stack : undefined,
+        installedRuleCount: 0,
+        packagesCount: 0,
+      };
+    }
+
+    let hasErrors = false;
+    const errorMessages: string[] = [];
+    let firstErrorStack: string | undefined;
+    let installedRuleCount = 0;
+
+    for (const [name, source] of Object.entries(expandedRules)) {
+      const ruleType = detectRuleType(source);
+      const ruleBasePath = getRuleSource(config, name);
+      const originalPresetPath = getOriginalPresetPath(config, name);
+
+      try {
+        let ruleContent;
+        switch (ruleType) {
+          case "npm":
+            ruleContent = collectNpmRule(name, source);
+            break;
+          case "local":
+            ruleContent = collectLocalRule(name, source, ruleBasePath);
+            break;
+          default:
+            errorMessages.push(`Unknown rule type: ${ruleType}`);
+            continue;
+        }
+
+        if (originalPresetPath) {
+          ruleContent.presetPath = originalPresetPath;
+        }
+
+        addRuleToCollection(ruleCollection, ruleContent, config.ides);
+        installedRuleCount++;
+      } catch (e) {
+        hasErrors = true;
+        const errorMessage = `Error processing rule ${name}: ${e instanceof Error ? e.message : String(e)}`;
+        errorMessages.push(errorMessage);
+
+        // Keep the first error stack trace
+        if (!firstErrorStack && e instanceof Error && e.stack) {
+          firstErrorStack = e.stack;
+        }
+      }
+    }
+
+    if (hasErrors) {
+      return {
+        success: false,
+        error: errorMessages.join("; "),
+        errorStack: firstErrorStack,
+        installedRuleCount,
+        packagesCount: 0,
+      };
+    }
+
+    writeRulesToTargets(ruleCollection);
+
+    if (config.mcpServers) {
+      const filteredMcpServers = Object.fromEntries(
+        Object.entries(config.mcpServers).filter(([, v]) => v !== false),
+      );
+      writeMcpServersToTargets(filteredMcpServers, config.ides, cwd);
+    }
+
+    return {
+      success: true,
+      installedRuleCount,
+      packagesCount: 1,
+    };
+  });
+}
